feat: fake site

+++ entrypoint.sh
@@ -1,5 +1,16 @@
 #!/bin/sh
 
+# 检查必要的环境变量
+if [ -z "$TROJAN_GO_SERVICE_PASSWORD" ]; then
+    echo "ENV TROJAN_GO_SERVICE_PASSWORD is required!"
+    exit 1
+fi
+
+if [ -z "$TROJAN_GO_SERVICE_DOMAIN" ]; then
+    echo "ENV TROJAN_GO_SERVICE_DOMAIN is required!"
+    exit 1
+fi
+
 # 1. 判断 Trojan-Go 配置是否存在
 # 1.1 配置文件 /etc/trojan-go/config.json
 rm -rf /etc/trojan-go/config.json
@@ -36,10 +47,19 @@ if [ ! -f /etc/caddy/Caddyfile ]; then
     exit 1
 fi
 
-# 3. 启动 Caddy
+# 3. 替换 Umami 的环境变量
+if [ -f /var/www/html/index.html ]; then
+    envsubst '${UMAMI_SRC} ${UMAMI_WEBSITE_ID} ${TURNSTILE_SITEKEY}' < /var/www/html/index.html > /var/www/html/index.html.tmp && \
+    mv /var/www/html/index.html.tmp /var/www/html/index.html
+else
+    echo "index.html not found in /var/www/html/"
+    exit 1
+fi
+
+# 4. 启动 Caddy
 caddy start --config /etc/caddy/Caddyfile
 
-# 4. 替换 Trojan-Go 配置文件内容
+# 5. 替换 Trojan-Go 配置文件内容
 # 替换域名和密码
 sed -i "s/your_password/$TROJAN_GO_SERVICE_PASSWORD/" /etc/trojan-go/config.json
 sed -i "s/your-domain-name.com/$TROJAN_GO_SERVICE_DOMAIN/" /etc/trojan-go/config.json
@@ -48,5 +68,5 @@ sed -i "s/your_cert.crt/\/etc\/trojan-go\/certificate.crt/" /etc/trojan-go/confi
 # 将 your_key.key 替换为 /etc/trojan-go/private.key
 sed -i "s/your_key.key/\/etc\/trojan-go\/private.key/" /etc/trojan-go/config.json
 
-# 5. 启动 Trojan-Go
+# 6. 启动 Trojan-Go
 /usr/local/bin/trojan-go -config /etc/trojan-go/config.json
